+++ src/is/_is.js
@@ -1,6 +1,7 @@
 
 // exporting modules to be included the UMD bundle
 
+import activeElement from './active-element';
 import disabled from './disabled';
 import focusRelevant from './focus-relevant';
 import focusable from './focusable';
@@ -11,6 +12,7 @@ import validArea from './valid-area';
 import validTabindex from './valid-tabindex';
 import visible from './visible';
 export default {
+  activeElement,
   disabled,
   focusRelevant,
   focusable,