+++ src/query/focusable.strict.js
@@ -4,51 +4,45 @@
 
 import isFocusable from '../is/focusable';
 import isFocusRelevant from '../is/focus-relevant';
-import isOnlyTabbable from '../is/only-tabbable';
 import getDocument from '../util/get-document';
 
-// see https://developer.mozilla.org/en-US/docs/Web/API/NodeFilter
-const FocusableFilter = function(node) {
-  if (node.shadowRoot) {
-    // return ShadowRoot elements regardless of them being focusable,
-    // so they can be walked recursively later
-    return NodeFilter.FILTER_ACCEPT;
-  }
+function createFilter(condition) {
+  // see https://developer.mozilla.org/en-US/docs/Web/API/NodeFilter
+  const filter = function(node) {
+    if (node.shadowRoot) {
+      // return ShadowRoot elements regardless of them being focusable,
+      // so they can be walked recursively later
+      return NodeFilter.FILTER_ACCEPT;
+    }
 
-  if (isFocusable(node)) {
-    // finds elements that could have been found by document.querySelectorAll()
-    return NodeFilter.FILTER_ACCEPT;
-  }
+    if (condition(node)) {
+      // finds elements that could have been found by document.querySelectorAll()
+      return NodeFilter.FILTER_ACCEPT;
+    }
 
-  return NodeFilter.FILTER_SKIP;
-};
-// IE requires a function, Browsers require {acceptNode: function}
-// see http://www.bennadel.com/blog/2607-finding-html-comment-nodes-in-the-dom-using-treewalker.htm
-FocusableFilter.acceptNode = FocusableFilter;
+    return NodeFilter.FILTER_SKIP;
+  };
+  // IE requires a function, Browsers require {acceptNode: function}
+  // see http://www.bennadel.com/blog/2607-finding-html-comment-nodes-in-the-dom-using-treewalker.htm
+  filter.acceptNode = filter;
+  return filter;
+}
 
-// see https://developer.mozilla.org/en-US/docs/Web/API/NodeFilter
-const PossiblyFocusableFilter = function(node) {
-  if (node.shadowRoot) {
-    // return ShadowRoot elements regardless of them being focusable,
-    // so they can be walked recursively later
-    return NodeFilter.FILTER_ACCEPT;
-  }
+const PossiblyFocusableFilter = createFilter(isFocusRelevant);
 
-  if (isOnlyTabbable(node) || isFocusRelevant(node)) {
-    // finds elements that could have been found by document.querySelectorAll()
-    return NodeFilter.FILTER_ACCEPT;
-  }
-
-  return NodeFilter.FILTER_SKIP;
-};
-// IE requires a function, Browsers require {acceptNode: function}
-// see http://www.bennadel.com/blog/2607-finding-html-comment-nodes-in-the-dom-using-treewalker.htm
-PossiblyFocusableFilter.acceptNode = PossiblyFocusableFilter;
-
-export default function queryFocusableStrict({context, includeContext, strategy} = {}) {
+export default function queryFocusableStrict({
+  context,
+  includeContext,
+  includeOnlyTabbable,
+  strategy,
+} = {}) {
   if (!context) {
     context = document.documentElement;
   }
+
+  const _isFocusable = isFocusable.rules.except({
+    onlyTabbable: includeOnlyTabbable,
+  });
 
   const _document = getDocument(context);
   // see https://developer.mozilla.org/en-US/docs/Web/API/Document/createTreeWalker
@@ -58,7 +52,7 @@ export default function queryFocusableStrict({context, includeContext, strategy}
     // element type filter
     NodeFilter.SHOW_ELEMENT,
     // custom NodeFilter filter
-    strategy === 'all' ? PossiblyFocusableFilter : FocusableFilter,
+    strategy === 'all' ? PossiblyFocusableFilter : createFilter(_isFocusable),
     // deprecated, but IE requires it
     false
   );
@@ -67,12 +61,13 @@ export default function queryFocusableStrict({context, includeContext, strategy}
 
   while (walker.nextNode()) {
     if (walker.currentNode.shadowRoot) {
-      if (isFocusable(walker.currentNode)) {
+      if (_isFocusable(walker.currentNode)) {
         list.push(walker.currentNode);
       }
 
       list = list.concat(queryFocusableStrict({
         context: walker.currentNode.shadowRoot,
+        includeOnlyTabbable,
         strategy,
       }));
     } else {
@@ -83,10 +78,10 @@ export default function queryFocusableStrict({context, includeContext, strategy}
   // add context if requested and focusable
   if (includeContext) {
     if (strategy === 'all') {
-      if (isOnlyTabbable(context) || isFocusRelevant(context)) {
+      if (isFocusRelevant(context)) {
         list.unshift(context);
       }
-    } else if (isFocusable(context)) {
+    } else if (_isFocusable(context)) {
       list.unshift(context);
     }
   }