+++ test/helper/supports.js
@@ -45,9 +45,12 @@ define(function(require) {
     canFocusScrollBody: require('ally/supports/focus-scroll-body')(),
     canFocusScrollContainer: require('ally/supports/focus-scroll-container')(),
     canFocusScrollContainerWithoutOverflow: require('ally/supports/focus-scroll-container-without-overflow')(),
-    canFocusSvgMethod: require('ally/supports/svg-focus-method')(),
+    canFocusSvg: require('ally/supports/focus-svg')(),
+    canFocusSvgFocusableAttribute: require('ally/supports/focus-svg-focusable-attribute')(),
+    canFocusSvgTabindexAttribute: require('ally/supports/focus-svg-tabindex-attribute')(),
     canFocusoutEvent: require('ally/supports/focusout-event')(),
     cssShadowPiercingDeepCombinator: require('ally/supports/css-shadow-piercing-deep-combinator')(),
+    svgFocusMethod: require('ally/supports/svg-focus-method')(),
     tabsequenceSortsAreaAtImagePosition: require('ally/supports/tabsequence-area-at-img-position')(),
   };