+++ test/unit/is.focus-relevant.test.js
@@ -31,6 +31,13 @@ define([
           isFocusRelevant(null);
         }).to.throw(TypeError, 'is/focus-relevant requires an argument of type Element');
       },
+      '.rules() and .except()': function() {
+        var element = document.getElementById('inert-div');
+        expect(isFocusRelevant.rules({
+          context: element,
+        })).to.equal(false, '.rules()');
+        expect(isFocusRelevant.rules.except({})(element)).to.equal(false, '.rules.except()');
+      },
       'inert div': function() {
         var element = document.getElementById('inert-div');
         expect(isFocusRelevant(element)).to.equal(false);
@@ -89,12 +96,12 @@ define([
       },
       'img with usemap': function() {
         var element = document.getElementById('img-usemap');
-        expect(isFocusRelevant(element)).to.equal(false);
+        expect(isFocusRelevant(element)).to.equal(supports.canFocusRedirectImgUsemap);
       },
       'img with usemap and tabindex': function() {
         var element = document.getElementById('img-usemap');
         element.setAttribute('tabindex', '-1');
-        expect(isFocusRelevant(element)).to.equal(supports.canFocusImgUsemapTabindex);
+        expect(isFocusRelevant(element)).to.equal(supports.canFocusRedirectImgUsemap || supports.canFocusImgUsemapTabindex);
       },
       'object element referencing svg': function() {
         var element = document.getElementById('object-svg');
@@ -127,12 +134,12 @@ define([
       },
       'label element': function() {
         var element = document.getElementById('label');
-        expect(isFocusRelevant(element)).to.equal(false);
+        expect(isFocusRelevant(element)).to.equal(true);
       },
       'label element with tabindex="-1"': function() {
         var element = document.getElementById('label');
         element.setAttribute('tabindex', '-1');
-        expect(isFocusRelevant(element)).to.equal(supports.canFocusLabelTabindex);
+        expect(isFocusRelevant(element)).to.equal(true);
       },
       'audio element': function() {
         var element = document.getElementById('audio');
@@ -144,16 +151,16 @@ define([
       },
       'svg element': function() {
         var element = document.getElementById('svg');
-        expect(isFocusRelevant(element)).to.equal(false);
+        expect(isFocusRelevant(element)).to.equal(supports.canFocusSvg);
       },
       'svg element with tabindex="-1"': function() {
         var element = document.getElementById('svg');
         element.setAttribute('tabindex', '-1');
-        expect(isFocusRelevant(element)).to.equal(supports.canFocusSvgMethod);
+        expect(isFocusRelevant(element)).to.equal(supports.canFocusSvg || supports.canFocusSvgTabindexAttribute);
       },
       'svg link element': function() {
         var element = document.getElementById('svg-link');
-        expect(isFocusRelevant(element)).to.equal(supports.canFocusSvgMethod);
+        expect(isFocusRelevant(element)).to.equal(true);
       },
       'embed element': function() {
         var element = document.getElementById('embed');